feat(wiki): adds popover and tooltip for Wiki
