+++ docs/zkEVM/index.md
@@ -12,15 +12,16 @@ keywords:
 image: https://wiki.polygon.technology/img/thumbnail/polygon-zkevm.png
 ---
 
-import { Welcome, Glossary } from "./landing.js";
+import Welcome from "./landing.js";
+import Glossary from "@theme/Popover";
 
 :::info Before We Begin
-
-<Glossary var="some_text"/>
 
 Polygon zkEVM Mainnet Beta is provided on an **AS-IS** and **AS-AVAILABLE** basis. The documentation contains statements about technical specifications, some of which may relate to future versions of Polygon zkEVM rather than its current implementation. Please read the [<ins>zkEVM Risk Disclosures</ins>](#polygon-zkevm-risk-disclosures) section.
 
 :::
+
+
 
 <Welcome/>
 
@@ -28,9 +29,7 @@ Polygon zkEVM Mainnet Beta is provided on an **AS-IS** and **AS-AVAILABLE** basi
 
 ## Polygon zkEVM Status
 
-<Glossary text="Some text" definition="And here goes the definition"/>
-
-You can check the live status of Polygon zkEVM network using the following links for Mainnet Beta and Public Testnet. <Glossary text="Some text" definition="And here goes the definition And here goes the definition And here goes the definition And here goes the definition And here goes the definition And here goes the definition And here goes the definition"/>
+You can check the live status of Polygon zkEVM network using the following links for Mainnet Beta and Public Testnet. <Glossary type="popover" text="Some text" definition="Here is the definition."/>
 
 - **Mainnet Beta &rarr;** https://status.zkevm-rpc.com
 - **Public Testnet &rarr;** https://status.zkevm-test.net
